feat: zod 관련 로직을 packages/types로 이동
