+++ apps/api/src/ticle/ticle.controller.ts
@@ -1,10 +1,11 @@
 import { Body, Controller, Get, Param, Post, Query, UseGuards, UsePipes } from '@nestjs/common';
+import { CreateTicleSchema } from '@repo/types';
 
 import { JwtAuthGuard } from '@/auth/jwt/jwt-auth.guard';
 import { GetUserId } from '@/common/decorator/get-userId.decorator';
 import { ZodValidationPipe } from '@/zodpipevalidation';
 
-import { CreateTicleDto, CreateTicleSchema } from './dto/createTicleDto';
+import { CreateTicleDto } from './dto/createTicleDto';
 import { GetTicleListQueryDto } from './dto/getTicleListQueryDto';
 import { TickleDetailResponseDto } from './dto/ticleDetailDto';
 import { SortType } from './sortType.enum';
